refactor: color-scheme and container-color in foundations
