+++ packages/foundations/scss/colors/classes/all.scss
@@ -3,12 +3,10 @@
 
 @each $variant in index.$variants {
 	@include get-class.init($variant);
-	@include get-class.init-container-colors($variant);
 }
 
 @each $additional-variant in index.$additional-variants {
 	@include get-class.init($additional-variant);
-	@include get-class.init-container-colors($additional-variant);
 }
 
 @each $background-color in index.$background-colors {
