+++ packages/foundations/scss/colors/classes/blue.scss
@@ -4,4 +4,3 @@
 $variant: "blue";
 
 @include get-class.init($variant);
-@include get-class.init-container-colors($variant);
