+++ packages/foundations/scss/colors/classes/brand.scss
@@ -4,4 +4,3 @@
 $variant: "brand";
 
 @include get-class.init($variant);
-@include get-class.init-container-colors($variant);
