+++ packages/foundations/scss/colors/classes/critical.scss
@@ -4,4 +4,3 @@
 $variant: "critical";
 
 @include get-class.init($variant);
-@include get-class.init-container-colors($variant);
