+++ packages/foundations/scss/colors/classes/cyan.scss
@@ -4,4 +4,3 @@
 $variant: "cyan";
 
 @include get-class.init($variant);
-@include get-class.init-container-colors($variant);
