+++ packages/foundations/scss/colors/classes/green.scss
@@ -4,4 +4,3 @@
 $variant: "green";
 
 @include get-class.init($variant);
-@include get-class.init-container-colors($variant);
