+++ packages/foundations/scss/colors/classes/informational.scss
@@ -4,4 +4,3 @@
 $variant: "informational";
 
 @include get-class.init($variant);
-@include get-class.init-container-colors($variant);
