+++ packages/foundations/scss/colors/classes/neutral.scss
@@ -4,4 +4,3 @@
 $variant: "neutral";
 
 @include get-class.init($variant);
-@include get-class.init-container-colors($variant);
