+++ packages/foundations/scss/colors/classes/orange.scss
@@ -4,4 +4,3 @@
 $variant: "orange";
 
 @include get-class.init($variant);
-@include get-class.init-container-colors($variant);
