+++ packages/foundations/scss/colors/classes/pink.scss
@@ -4,4 +4,3 @@
 $variant: "pink";
 
 @include get-class.init($variant);
-@include get-class.init-container-colors($variant);
