+++ packages/foundations/scss/colors/classes/red.scss
@@ -4,4 +4,3 @@
 $variant: "red";
 
 @include get-class.init($variant);
-@include get-class.init-container-colors($variant);
