+++ packages/foundations/scss/colors/classes/successful.scss
@@ -4,4 +4,3 @@
 $variant: "successful";
 
 @include get-class.init($variant);
-@include get-class.init-container-colors($variant);
