+++ packages/foundations/scss/colors/classes/turquoise.scss
@@ -4,4 +4,3 @@
 $variant: "turquoise";
 
 @include get-class.init($variant);
-@include get-class.init-container-colors($variant);
