+++ packages/foundations/scss/colors/classes/violet.scss
@@ -4,4 +4,3 @@
 $variant: "violet";
 
 @include get-class.init($variant);
-@include get-class.init-container-colors($variant);
