+++ packages/foundations/scss/colors/classes/warning.scss
@@ -4,4 +4,3 @@
 $variant: "warning";
 
 @include get-class.init($variant);
-@include get-class.init-container-colors($variant);
