+++ packages/foundations/scss/colors/classes/yellow.scss
@@ -4,4 +4,3 @@
 $variant: "yellow";
 
 @include get-class.init($variant);
-@include get-class.init-container-colors($variant);
