+++ packages/migration/src/migration/v0.0.7-v1.0.0.ts
@@ -1,0 +1,14 @@
+import type { ReplaceInFileConfig } from 'replace-in-file';
+
+export const v007_v100: ReplaceInFileConfig[] = [
+	{
+		files: '',
+		from: /container-color/g,
+		to: 'color'
+	},
+	{
+		files: '',
+		from: /color-scheme/g,
+		to: 'mode'
+	}
+];
