+++ packages/migration/test/v0.0.7-v1.0.0/has-changes.txt
@@ -1,0 +1,3 @@
+class="db-container-color-neutral"
+[data-container-color]="neutral"
+[data-color-scheme]="dark"
