+++ packages/migration/test/v0.0.7-v1.0.0/no-change.txt
@@ -1,0 +1,3 @@
+class="db-color-neutral"
+[data-color]="neutral"
+[data-mode]="dark"
